Make bootloader start a bit safer, adds 34 bytes
Risk scenario for a worst case buggy/malicious application:
-Request bootloader start (8F 31)
-Continuously request reset (02 00)
This causes bootloaderTimer to never decrement, and bootloaderFlag reset
to be set. Causing that if the user press Power long time after, the
application jumps to bootloader.

This new implementation does a few things:
-Keypress will only decrement countdown to 0.5 second, it will not keep or
increment the countdown. Ensuring that this will never be active more than
20 seconds.
-Require that power and reset buttons are actually pressed down. Ignore I2C
and keyboard triggers. Also require NMI to NOT be pressed.
-Evaluate keypress in mainloop while decrementing counter

Code size: 6676 bytes