GPIO library: Add dedicated inputWithPullup/driveLow, for PS2
This is to simplify toggling between the two meaningful states: Input with
pullup, and output driving low.

This is particularly useful for PS2, which is open collector logic,
requiring a strong pullup. The external 4.7k pullup is not always strong
enough, but together with the 20-50k additional internal pullup, it handles
all known keyboards.